refactor: move public types to their own file
BREAKING CHANGE: NMRium is now a named export and some types were renamed for consistency
